How to avoid 15s delay from SC command.
From the SkySafari manual, quote:

Set Time and Location: If turned on, SkySafari will send
the time and location from your iPhone to the telescope
when establishing a connection. This will overwrite your
telescope's previously-set time and location. For older
Meade LX-200 telescopes, this may also cause a delay of
up to 15 seconds when connecting, so you may wish to turn
this setting off.