Record local site position & date/time.
Currently using astropysics Site class for location and timezone.
Will probably migrate this to using AstroPy in future, if and when
that gains support for observation sites and conversion between
local horizon coordinates (Alt/Az) and equatorial coordinates
(RA/Dec).

Storing desired site data/time as an offset (in seconds) from the
system clock. This will be particularly useful on systems like a
Raspberry Pi which lack a clock battery, and therefore a meaningful
date if run offline.