Connect GY-80 orientation sensor to telescope server
This 'works' in that I can move the sensor by hand and see the
reticule move in SkySafari Plus v4.0 running on an iPad. However,
the location signal is very noisy and therefore wobbles about -
smoothing and/or using the gyro will be required.

Assuming the telescope is aimed along the X axis of the sensor,
the basic orientation/mapping to Alt/Az seems about right.
Wiyll also need alignment support - simplest approach is use this
to update the magnetic deviation and correct the azimuth angle.