Use simple complementatry filter (2% acc/mag, 98% gyro).
Still not sure what the intent was in the code I was reading from
https://github.com/DavidEGrayson/minimu9-ahrs/blob/master/minimu9-ahrs.cpp
and http://sourceforge.net/p/firetail/code/ci/master/tree/ahrs.cpp
but my attempt to implement it would just keep drifting.

This works; the combined result lags the immediate acc/mag only
output, but does seem to wobble about much less thanks to the gyro.
Even so, it is annoyingly jumpy in indoor testing feeding this to
SkySafari - so more smoothing may be required...