Support basic NexStar commands (V, E, e).
This is enough for Sky Safari Plus v4.0.1 to connect as a push-to
mount (it sends 'V' to get the version), and track the position
(it sends only 'e' for high precision, also implemented 'E' for
low precision). See:

http://www.celestron.com/c3/images/files/downloads/1154108406_nexstarcommprot.pdf

Interestlying, unlike with the Meade LX200 protocol, Sky Safari
seems to do alignment internally (but only allows a nearby star
to be used). i.e. It does not send the 'S' command.

Also Sky Safari does not attempt to set the local site location
(via the 'W' command), nor time (via the 'H' command).